Refiner modifications and items for REEs
Added H2SO4 recipe (and items for making it) and modified the Refiner to use it when "acid" is present in a recipe file name. For now no recipes use it. Future REEs refining will use it, and maybe PGMs.
Modded coals have been fixed in previous commit (campfire & fire charge)