Adjusted OreGeneration and created OreUtilities
OreGeneration now utilizes OreUtilities lists and functions to generate ores, with OreUtilities containing various config values for each ore as well as the grouping of ores into specific categories. OreGeneration is more simplified as a result, and OreUtilities has the potential to be utilized in other ways in the future.